chore: housekeeping — lint fixes, file organization, gitignore cleanup
- Fix clippy error in server/src/lib.rs: remove absurd `port > 65535`
  comparison (always false for u16)
- Fix clippy warnings: single_match → if let in tui/src/input.rs,
  redundant closure in tui/src/events.rs, unnecessary u32 cast in
  agent/src/workflow.rs, &PathBuf → &Path in src/main.rs functions,
  print_literal in src/main.rs
- Untrack .nodestradamus/*.msgpack binary cache files; add to .gitignore
- Move pipeline data files (*.jsonl) to data/; add data/ to .gitignore
- Delete stale root files: package-lock.json, s1/s2/s3_status.txt,
  STRESS_TEST_FINAL_REPORT.txt
- Add scripts/spans_to_annotations.py and stress_50_prompts.txt
- Reduce default_guaranteed_recent_turns from 30 → 4 (agent/context.rs)
- Add test_empty_response and test_streaming_roundtrip docs (llm/tests)

+++ crates/agent/src/workflow.rs
@@ -629,7 +629,7 @@ pub async fn run_agent_loop(
         all_node_ids.extend(tool_result_ids.iter().cloned());
 
         // Check for soft escalation after tools execute
-        if check_soft_escalation(turn as u32, &session.agent_config, &response, events) {
+        if check_soft_escalation(turn, &session.agent_config, &response, events) {
             // Agent should respond to the escalation by synthesizing findings.
             // The synthesis directive is in the SoftEscalationTriggered event.
             // For now, we continue the loop so the agent can respond with synthesis.
